+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.0]
+### Added
+- `limit` parameter for `api.changeset.get_query()` endpoint.
+
+### Fixed
+- Corrected character when adding parameters in endpoint `api.changeset.get_query()` (from `;` to `&`).
+
 ## [1.0.2]
 ### Fixed
 - Auth problems in API when using characters unsupported by latin-1 codec.